moved packages: csds-client; cleanup
