+++ packages/csds-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lp-client/csds-config",
+  "name": "@lp-libs/csds-config",
   "version": "0.1.0",
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
