+++ packages/csds-resolver/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lp-client/csds-resolver",
+  "name": "@lp-libs/csds-resolver",
   "version": "0.1.0",
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
@@ -19,7 +19,7 @@
     "prepare": "tsc -p tsconfig.build.json"
   },
   "dependencies": {
-    "@lp-client/csds-config": "^0.1.0"
+    "@lp-libs/csds-config": "^0.1.0"
   },
   "devDependencies": {
     "typescript": "^2.7.2"
