+++ packages/logger-console/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lp-client/logger-console",
+  "name": "@lp-libs/logger-console",
   "version": "0.1.0",
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
@@ -19,7 +19,7 @@
     "prepare": "tsc -p tsconfig.build.json"
   },
   "dependencies": {
-    "@lp-client/logger": "^0.1.0",
+    "@lp-libs/logger": "^0.1.0",
     "chalk": "^2.3.0"
   },
   "devDependencies": {
